+++ demos/ember-app/app/controllers/application.js
@@ -1,8 +1,0 @@
-import Controller from '@ember/controller';
-import { camelize } from '@esri/hub-common';
-
-export default class ApplicationController extends Controller {
-  get message () {
-    return camelize('Hello  Ember');
-  }
-}